migrate /api/book/confirm to viewer.bookings.confirm
