+++ apps/web/components/booking/BookingListItem.tsx
@@ -1,7 +1,6 @@
 import { BookingStatus } from "@prisma/client";
 import { useRouter } from "next/router";
 import { useState } from "react";
-import { useMutation } from "react-query";
 
 import dayjs from "@calcom/dayjs";
 import classNames from "@calcom/lib/classNames";
@@ -15,7 +14,6 @@ import { Icon } from "@calcom/ui/Icon";
 import { Tooltip } from "@calcom/ui/Tooltip";
 import { TextArea } from "@calcom/ui/form/fields";
 
-import { HttpError } from "@lib/core/http/error";
 import useMeQuery from "@lib/hooks/useMeQuery";
 import { linkValueToString } from "@lib/linkValueToString";
 import { LocationType } from "@lib/location";
@@ -43,39 +41,29 @@ function BookingListItem(booking: BookingItemProps) {
   const router = useRouter();
   const [rejectionReason, setRejectionReason] = useState<string>("");
   const [rejectionDialogIsOpen, setRejectionDialogIsOpen] = useState(false);
-  const mutation = useMutation(
-    async (confirm: boolean) => {
-      let body = {
-        id: booking.id,
-        confirmed: confirm,
-        language: i18n.language,
-        reason: rejectionReason,
-      };
-      /**
-       * Only pass down the recurring event id when we need to confirm the entire series, which happens in
-       * the "Recurring" tab, to support confirming discretionally in the "Recurring" tab.
-       */
-      if (booking.listingStatus === "recurring" && booking.recurringEventId !== null) {
-        body = Object.assign({}, body, { recurringEventId: booking.recurringEventId });
-      }
-      const res = await fetch("/api/book/confirm", {
-        method: "PATCH",
-        body: JSON.stringify(body),
-        headers: {
-          "Content-Type": "application/json",
-        },
-      });
-      if (!res.ok) {
-        throw new HttpError({ statusCode: res.status });
-      }
+  const mutation = trpc.useMutation(["viewer.bookings.confirm"], {
+    onSuccess: () => {
       setRejectionDialogIsOpen(false);
+      utils.invalidateQueries("viewer.bookings");
     },
-    {
-      async onSettled() {
-        await utils.invalidateQueries(["viewer.bookings"]);
-      },
+  });
+
+  const bookingConfirm = async (confirm: boolean) => {
+    let body = {
+      bookingId: booking.id,
+      confirmed: confirm,
+      reason: rejectionReason,
+    };
+    /**
+     * Only pass down the recurring event id when we need to confirm the entire series, which happens in
+     * the "Recurring" tab, to support confirming discretionally in the "Recurring" tab.
+     */
+    if (booking.listingStatus === "recurring" && booking.recurringEventId !== null) {
+      body = Object.assign({}, body, { recurringEventId: booking.recurringEventId });
     }
-  );
+    mutation.mutate(body);
+  };
+
   const isUpcoming = new Date(booking.endTime) >= new Date();
   const isCancelled = booking.status === BookingStatus.CANCELLED;
   const isConfirmed = booking.status === BookingStatus.ACCEPTED;
@@ -102,7 +90,7 @@ function BookingListItem(booking: BookingItemProps) {
           ? t("confirm_all")
           : t("confirm"),
       onClick: () => {
-        mutation.mutate(true);
+        bookingConfirm(true);
       },
       icon: Icon.FiCheck,
       disabled: mutation.isLoading,
@@ -287,7 +275,7 @@ function BookingListItem(booking: BookingItemProps) {
             <Button
               disabled={mutation.isLoading}
               onClick={() => {
-                mutation.mutate(false);
+                bookingConfirm(false);
               }}>
               {t("rejection_confirmation")}
             </Button>
