+++ packages/prisma/zod-utils.ts
@@ -87,6 +87,13 @@ export const bookingCreateBodySchema = z.object({
 
 export type BookingCreateBody = z.input<typeof bookingCreateBodySchema>;
 
+export const bookingConfirmPatchBodySchema = z.object({
+  bookingId: z.number(),
+  confirmed: z.boolean(),
+  recurringEventId: z.string().optional(),
+  reason: z.string().optional(),
+});
+
 export const extendedBookingCreateBody = bookingCreateBodySchema.merge(
   z.object({
     noEmail: z.boolean().optional(),
